domain  cov 100% and reviewed

+++ domain/entities/car.py
@@ -1,9 +1,10 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass(frozen=True)
 class Car:
     make: str
     model: str
-    value: float
+    value: Decimal
     year: int
