+++ domain/events/premium_calculated.py
@@ -1,12 +1,13 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 from domain.entities.car import Car
 
 
 @dataclass(frozen=True)
 class PremiumCalculated:
-    applied_rate: float
-    calculated_premium: float
+    applied_rate: Decimal
+    calculated_premium: Decimal
     car: Car
-    deductible_value: float
-    policy_limit: float
+    deductible_value: Decimal
+    policy_limit: Decimal
