+++ domain/value_objects/policy_limit_breakdown.py
@@ -1,8 +1,9 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass(frozen=True)
 class PolicyLimitBreakdown:
-    base_policy_limit: float
-    deductible_value: float
-    policy_limit: float
+    base_policy_limit: Decimal
+    deductible_value: Decimal
+    policy_limit: Decimal
