+++ domain/value_objects/premium_breakdown.py
@@ -1,9 +1,10 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass(frozen=True)
 class PremiumBreakdown:
-    applied_rate: float
-    base_premium: float
-    calculated_premium: float
-    deductible_discount: float
+    applied_rate: Decimal
+    base_premium: Decimal
+    calculated_premium: Decimal
+    deductible_discount: Decimal
